Moved setup and tear down to one main file
